Adding customFilePath parameter.
This commit proposes the addition of a "customPaperFilePath" parameter taking a path to a .bin file for TD and RJ type printers (Only tested on a TD-2120N), for example:

thePrinter.customPaperFilePath = 'assets / printer / bin / 51x26.bin';

The addition concerns the Android and iOS platforms.

I don't know if the code is optimal, maybe a review is needed.